Graph core superclass
